Divide configureStore into .dev, .prod / Add redux-logger as dev-dependency / Enable Redux-ChromeDevtool / Add REACT_EDITOR on .env.sample
