+++ tilup-web/.env.sample
@@ -1,2 +1,3 @@
 REACT_APP_API_HOST="http://localhost:3000"
 PORT=3005
+REACT_EDITOR=atom
