+++ tilup-web/src/store/configureStore.js
@@ -1,12 +1,5 @@
-import { createStore, applyMiddleware, compose } from 'redux';
-import rootReducer from '../reducers';
-import ReduxThunk from 'redux-thunk';
-
-const createAppStore = compose(
-	applyMiddleware(ReduxThunk)
-)(createStore);
-
-export default function configureStore(initialState){
-	const store = createAppStore(rootReducer, initialState);
-	return store;
+if ( process.env.NODE_ENV === 'production' ) {
+  module.exports = require('./configureStore.prod');
+} else {
+  module.exports = require('./configureStore.dev');
 }
